SERVICES-1502: move wrapping services to own folder
