+++ src/modules/wrapping/wrap.resolver.ts
@@ -3,8 +3,8 @@ import { Inject, UseGuards } from '@nestjs/common';
 import { TransactionModel } from '../../models/transaction.model';
 import { EsdtToken } from 'src/modules/tokens/models/esdtToken.model';
 import { WrapModel } from './models/wrapping.model';
-import { WrapService } from './wrap.service';
-import { TransactionsWrapService } from './transactions-wrap.service';
+import { WrapService } from './services/wrap.service';
+import { TransactionsWrapService } from './services/wrap.transactions.service';
 import { ApolloError } from 'apollo-server-express';
 import { JwtOrNativeAuthGuard } from '../auth/jwt.or.native.auth.guard';
 import { AuthUser } from '../auth/auth.user';
